website/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
black==24.8.0
click==8.1.7
colorama==0.4.6
dj-database-url==2.2.0
Django==4.2.15
django-bootstrap-v5==1.0.11
django-dbbackup==4.2.1
django-environ==0.11.2
django-jquery==3.1.0
django-phonenumber-field==8.0.0
fontawesomefree==6.6.0
gunicorn==23.0.0
mssql-django==1.5
mypy-extensions==1.0.0
numpy==2.1.2
opencv-python-headless==4.10.0.84
packaging==24.1
pathspec==0.12.1
phonenumbers==8.13.45
phonenumberslite==8.13.45
pillow==10.4.0
platformdirs==4.2.2
psycopg2-binary==2.9.10
pyodbc==5.1.0
python-decouple==3.8
pytz==2024.2
pyzbar==0.1.9
soupsieve==2.5
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1